generate_lts_report: Add --force-good
Sometimes qa-reports reports a regression when there isn't actually any.
In such a case, use --force-good to force the "No regressions" header.
Note that the report will still need to be modified because it will show
'Regressions' inline.

To do this, switch to argparse for argument parsing. Also DRY up the
multiple places that branches were being listed. Lastly, drop
4.16.

Signed-off-by: Dan Rue <dan.rue@linaro.org>