Initial trigger script for developer notifications
TODO:
  - builds url and state file are hard coded.
  - doesn't actually email; just prints report
  - Perhaps limit the number of builds to consider. By date? To avoid
  flooding notifications in the event of losing the state file.
  - Retrieve email address from metadata

This script is designed to run regularly from jenkins. Each time it
runs, it will retrieve the list of builds from the developer project,
and for each build, if it has not yet been notified (based on data from
the state file), and if it is finished, it will send a notification.

Signed-off-by: Dan Rue <dan.rue@linaro.org>