Update to provide total at the end
Example output from today:

Ran 956645 tests on 54 builds on branch stable v4.4.y on OE in the last 86 days.
Ran 138608 tests on 47 builds on branch linaro-hikey-stable v4.4.y on OE in the last 86 days.
Ran 1327738 tests on 61 builds on branch stable v4.9.y on OE in the last 86 days.
Ran 1288546 tests on 58 builds on branch stable v.4.14.y on OE in the last 86 days.
Ran 1205498 tests on 59 builds on branch stable v4.19.y on OE in the last 86 days.
Ran 1099805 tests on 55 builds on branch stable v4.20.y on OE in the last 86 days.
Ran 318236 tests on 14 builds on branch stable v5.0.y on OE in the last 86 days.
Ran 3878324 tests on 189 builds on branch mainline on OE in the last 86 days.
Ran 946438 tests on 52 builds on branch next on OE in the last 86 days.
Ran 11159838 total tests on 589 builds in the last 86 days.

Signed-off-by: Dan Rue <dan.rue@linaro.org>