list_rc_log: Use GitPython to traverse public-inbox' collection of emails
Instead of fetching the file m from the Git repository using a
subprocess (which we always knew was inefficient), use
GitPython to directly go through all the commits available
while looking for review requests and replies.

We still need to make email objects out of the git commits,
but this is now done only on relevant commits, not all of
them. This saves *a lot* of processing time.

With this change, the objects stored in the data structure are
no longer email messages, but Git commits. Those Git commits
can easily be converted to email messages with the
`commit_to_email_message()` helper.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>