build_info: Add --incomplete and --save-logs options
These options help in understanding what's going on with all
incomplete jobs for a given build.

First, --incomplete will provide a list of incomplete jobs
and their reason for failure when running with --print-jobs.

Then, --save-logs will also fetch and save the LAVA logs of
those incomplete jobs to the current directory, making it easy
to grep for important information.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>